app_test.py
from flask import Flask, render_template, request

app = Flask(__name__)

def shutdown_server():
    func = request.environ.get('werkzeug.server.shutdown')
    if func is None:
        raise RuntimeError('Not running with the Werkzeug Server')
    func()

@app.route('/shutdown', methods=['POST'])
def shutdown():
    shutdown_server()
    return 'Server shutting down...'

@app.route('/')
def menu():
    return render_template('index.html')

def main():
    app.run(debug=True)

if __name__ == '__main__':
    main()